Issue 37739: GH 3424 Add SqlJson ToString()
- Added ToString() to SqlJson.
- Added missing commentary to public API docs.
- Simplified implementation.
- Improved tests and moved them to the UnitTests project.
